- trying "wait for file" to solve the windows crashes

+++ mole/qgisinteraction/layer_interaction.py
@@ -112,23 +112,18 @@ def remove_filegroup(path,filenamebase,ext=[],ignore=[]):
     import os
     os.environ['PATH'] += ":"+"/usr/local/bin"
     tgt_files=filter(lambda x: x.startswith(filenamebase+'.'),os.listdir(path))
-    #print tgt_files
     if ext:
         tgt_files=filter(lambda x: x.split('.')[1] in ext , tgt_files)
-    #print tgt_files
     if ignore:
         tgt_files=filter(lambda x: not x.split('.')[1]  in ignore , tgt_files)
-    #print tgt_files
     for i in tgt_files:
         if os.path.exists(os.path.join(path,i)):
             try:
-                #print i
                 os.remove(os.path.join(path,i))
-               # os.wait()
             except:
                 QgsMessageLog.logMessage("remove_filegroup : Could not remove '"+str(os.path.join(path,i))+"' !",'Error in remove_filegroup', QgsMessageLog.CRITICAL)
                 oeq_global.OeQ_init_error("remove_filegroup :","Could not remove '"+str(os.path.join(path,i))+"' !")
-                #return None
+                return None
     return tgt_files
 
 def rename_filegroup(path,filenamebase,newfilenamebase,ext=[],ignore=[]):
@@ -136,27 +131,19 @@ def rename_filegroup(path,filenamebase,newfilenamebase,ext=[],ignore=[]):
     import os
     os.environ['PATH'] += ":"+"/usr/local/bin"
     src_files=filter(lambda x: x.startswith(filenamebase+'.'),os.listdir(path))
-    #print src_files
     if ext:
         src_files=filter(lambda x: x.split('.')[1] in ext , src_files)
-    #print src_files
     if ignore:
         src_files=filter(lambda x: not x.split('.')[1]  in ignore , src_files)
-    #print src_files
-    #print [newfilenamebase+'.'+i.split('.')[1] for i in src_files]
     tgt_files=[[i,newfilenamebase+'.'+i.split('.')[1]] for i in src_files]
-    #print tgt_files
     for i in tgt_files:
-        #if os.path.exists(os.path.join(path,i[0])):
-            #try:
-                #print i[0]
-                #print i[1]
+        if os.path.exists(os.path.join(path,i[0])):
+            try:
                 os.rename(os.path.join(path,i[0]),os.path.join(path,i[1]))
-               # os.wait()
-            #except:
-               # QgsMessageLog.logMessage("rename_filegroup : Could not rename '"+str(os.path.join(path,i[0]))+"' to '"+str(os.path.join(path,i[1]))+"' !",'Error in rename_filegroup', QgsMessageLog.CRITICAL)
-                #oeq_global.OeQ_init_error("rename_filegroup :","Could not rename '"+str(os.path.join(path,i[0]))+"' to '"+str(os.path.join(path,i[1]))+"' !")
-                #return None
+            except:
+                QgsMessageLog.logMessage("rename_filegroup : Could not rename '"+str(os.path.join(path,i[0]))+"' to '"+str(os.path.join(path,i[1]))+"' !",'Error in rename_filegroup', QgsMessageLog.CRITICAL)
+                oeq_global.OeQ_init_error("rename_filegroup :","Could not rename '"+str(os.path.join(path,i[0]))+"' to '"+str(os.path.join(path,i[1]))+"' !")
+                return None
     return tgt_files
 
 def remove_layer(layer,physical=False):
